Move tests into package.
