+++ sparse/tests/test_coo.py
