+++ sparse/tests/test_io.py
